Refactoring
Engine is used as Pattern Engine for one pattern and is now instanced by GUI and by App
GUI can now fire change state Events.
App is now controlling the processing of LEDstrip and of GUI
RestAPI is to be fixed since the invoke-paths are not valid anymore.
Additional States are defined.

ToDo:
- Implementing new Unstable state
- optimizing appearance of all states
- fixing ApiController-Invokes

ToDo Later:
- Adding GUI to manipulate state visualization